Privatize thread-local storage headers
