+++ grpc.gemspec
@@ -65,10 +65,6 @@ Gem::Specification.new do |s|
   s.files += %w( include/grpc/support/sync_windows.h )
   s.files += %w( include/grpc/support/thd.h )
   s.files += %w( include/grpc/support/time.h )
-  s.files += %w( include/grpc/support/tls.h )
-  s.files += %w( include/grpc/support/tls_gcc.h )
-  s.files += %w( include/grpc/support/tls_msvc.h )
-  s.files += %w( include/grpc/support/tls_pthread.h )
   s.files += %w( include/grpc/support/useful.h )
   s.files += %w( include/grpc/impl/codegen/atm.h )
   s.files += %w( include/grpc/impl/codegen/atm_gcc_atomic.h )
@@ -93,6 +89,10 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/gpr/string_windows.h )
   s.files += %w( src/core/lib/gpr/thd_internal.h )
   s.files += %w( src/core/lib/gpr/time_precise.h )
+  s.files += %w( src/core/lib/gpr/tls.h )
+  s.files += %w( src/core/lib/gpr/tls_gcc.h )
+  s.files += %w( src/core/lib/gpr/tls_msvc.h )
+  s.files += %w( src/core/lib/gpr/tls_pthread.h )
   s.files += %w( src/core/lib/gpr/tmpfile.h )
   s.files += %w( src/core/lib/gprpp/abstract.h )
   s.files += %w( src/core/lib/gprpp/atomic.h )
