+++ include/grpc/module.modulemap
@@ -17,7 +17,6 @@ framework module grpc {
   header "support/sync_generic.h"
   header "support/thd.h"
   header "support/time.h"
-  header "support/tls.h"
   header "support/useful.h"
   header "impl/codegen/atm.h"
   header "impl/codegen/fork.h"
@@ -63,9 +62,6 @@ framework module grpc {
   textual header "support/sync_custom.h"
   textual header "support/sync_posix.h"
   textual header "support/sync_windows.h"
-  textual header "support/tls_gcc.h"
-  textual header "support/tls_msvc.h"
-  textual header "support/tls_pthread.h"
   textual header "impl/codegen/atm_gcc_atomic.h"
   textual header "impl/codegen/atm_gcc_sync.h"
   textual header "impl/codegen/atm_windows.h"
