+++ src/core/lib/gpr/tls_gcc.h
@@ -16,27 +16,27 @@
  *
  */
 
-#ifndef GRPC_SUPPORT_TLS_MSVC_H
-#define GRPC_SUPPORT_TLS_MSVC_H
+#ifndef GRPC_CORE_LIB_GPR_TLS_GCC_H
+#define GRPC_CORE_LIB_GPR_TLS_GCC_H
 
-/** Thread local storage based on ms visual c compiler primitives.
+#include <stdbool.h>
+
+#include <grpc/support/log.h>
+
+/** Thread local storage based on gcc compiler primitives.
    #include tls.h to use this - and see that file for documentation */
 
-struct gpr_msvc_thread_local {
+struct gpr_gcc_thread_local {
   intptr_t value;
 };
 
-/** Use GPR_TLS_DECL to declare tls static variables outside a class */
 #define GPR_TLS_DECL(name) \
-  static __declspec(thread) struct gpr_msvc_thread_local name = {0}
+  static __thread struct gpr_gcc_thread_local name = {0}
 
-/** Use GPR_TLS_CLASS_DECL to declare tls static variable members of a class.
- *  GPR_TLS_CLASS_DEF needs to be called to define this member. */
 #define GPR_TLS_CLASS_DECL(name) \
-  static __declspec(thread) struct gpr_msvc_thread_local name
+  static __thread struct gpr_gcc_thread_local name
 
-#define GPR_TLS_CLASS_DEF(name) \
-  __declspec(thread) struct gpr_msvc_thread_local name = {0}
+#define GPR_TLS_CLASS_DEF(name) __thread struct gpr_gcc_thread_local name = {0}
 
 #define gpr_tls_init(tls) \
   do {                    \
@@ -47,4 +47,4 @@ struct gpr_msvc_thread_local {
 #define gpr_tls_set(tls, new_value) (((tls)->value) = (new_value))
 #define gpr_tls_get(tls) ((tls)->value)
 
-#endif /* GRPC_SUPPORT_TLS_MSVC_H */
+#endif /* GRPC_CORE_LIB_GPR_TLS_GCC_H */
