+++ test/core/gpr/tls_test.cc
@@ -21,9 +21,10 @@
 #include <grpc/support/log.h>
 #include <grpc/support/sync.h>
 #include <grpc/support/thd.h>
-#include <grpc/support/tls.h>
 #include <stdio.h>
 #include <stdlib.h>
+
+#include "src/core/lib/gpr/tls.h"
 #include "test/core/util/test_config.h"
 
 #define NUM_THREADS 100
